requirements.txt
admin-multiupload==1.10
Django==2.1
django-cors-headers==2.4.0
django-filter==2.0.0
django-reset-migrations==0.3.1
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
Pillow==5.2.0
psycopg2==2.7.5
PyJWT==1.6.4
pytz==2018.5
requests==2.19.1